lib/bb: Drop BBFILE_PRIORITY support
These entries, usually in layer.conf are part of an old, pre-layers
part of bitbake for overlay/collection support. This is not really
used any more and was replaced by the idea of layers.

In a recent discussion, we asked the question "When do you need to
use proprities?" and the answer was "almost never" given all the other
tools and techniques now available in layers.

These layer priorities cause users huge amounts of confusion. It is
unclear to people how the "magic values" work if it is unset
and if they are set differently, users don't expect the way it
actually works and the effects it has between layers.

In discussions I've had, there has been surprisingly strong support
for removing them entirely, even if that does means we'd need to fix
some corner cases.

I wholly agree and support that idea, so lets remove the support
and take a simplification where we can. The functionality had a
time and place but it has passed and we've moved on.

Developers are about to need to change layer.conf to update the
series compatibility for the next release so it is a good time to
do it.

This patch removes the core support and replaces it with a warning.
It does not chase down every last reference or function parmeter,
that can follow as time allows.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>